added one more field 'Quantity' to ProductSize

+++ src/main/java/kg/demo/dodo/model/response/ProductSizeListResponse.java
@@ -12,5 +12,5 @@ public interface ProductSizeListResponse {
     String getDescription();
     String getCategory();
     String getSize();
-    BigDecimal getPrice();
+    Double getPrice();
 }
